+++ LogisticsHelpSystemLibrary/Models/Database/ApplicationDb/Component.cs
@@ -17,13 +17,13 @@ namespace LogisticsHelpSystemLibrary.Models.Database.ApplicationDb
         [Required]
         public string ComponentName { get; set; }
 
-        public double? ComponentWidth { get; set; }
+        public float? ComponentWidth { get; set; }
 
-        public double? ComponentHeigth { get; set; }
+        public float? ComponentHeigth { get; set; }
 
-        public double? ComponentDepth { get; set; }
+        public float? ComponentDepth { get; set; }
 
-        public double? ComponentWeigth { get; set; }
+        public float? ComponentWeigth { get; set; }
 
         public virtual ICollection<Order> Orders { get; set; }
     }
